+++ apps/astarte_appengine_api/lib/astarte_appengine_api/device/device_status.ex
@@ -20,7 +20,7 @@ defmodule Astarte.AppEngine.API.Device.DeviceStatus do
   import Ecto.Changeset
 
   alias Astarte.AppEngine.API.Device.DeviceStatus
-  alias Astarte.AppEngine.API.Device.InterfaceVersion
+  alias Astarte.AppEngine.API.Device.InterfaceInfo
   alias Astarte.Core.Device
 
   @primary_key {:id, :binary_id, autogenerate: false}
@@ -37,6 +37,7 @@ defmodule Astarte.AppEngine.API.Device.DeviceStatus do
     field :credentials_inhibited, :boolean
     field :total_received_msgs, :integer
     field :total_received_bytes, :integer
+    field :previous_interfaces, {:array, :map}
     field :groups, {:array, :string}
   end
 
@@ -62,12 +63,48 @@ defmodule Astarte.AppEngine.API.Device.DeviceStatus do
       "inhibit_credentials_request" => credentials_inhibited,
       "total_received_msgs" => total_received_msgs,
       "total_received_bytes" => total_received_bytes,
-      "groups" => groups_map
+      "groups" => groups_map,
+      "exchanged_msgs_by_interface" => exchanged_msgs_by_interface,
+      "exchanged_bytes_by_interface" => exchanged_bytes_by_interface,
+      "old_introspection" => old_introspection
     } = row
 
     introspection =
-      Map.merge(introspection_major || %{}, introspection_minor || %{}, fn _key, major, minor ->
-        %InterfaceVersion{major: major, minor: minor}
+      Map.merge(introspection_major || %{}, introspection_minor || %{}, fn
+        interface_name, major, minor ->
+          exchanged_msgs =
+            (exchanged_msgs_by_interface || %{})
+            |> Map.get({interface_name, major}, 0)
+
+          exchanged_bytes =
+            (exchanged_bytes_by_interface || %{})
+            |> Map.get({interface_name, major}, 0)
+
+          %InterfaceInfo{
+            major: major,
+            minor: minor,
+            exchanged_msgs: exchanged_msgs,
+            exchanged_bytes: exchanged_bytes
+          }
+      end)
+
+    previous_interfaces =
+      Enum.map(old_introspection || %{}, fn {{interface_name, major}, minor} ->
+        exchanged_msgs =
+          (exchanged_msgs_by_interface || %{})
+          |> Map.get({interface_name, major}, 0)
+
+        exchanged_bytes =
+          (exchanged_bytes_by_interface || %{})
+          |> Map.get({interface_name, major}, 0)
+
+        %InterfaceInfo{
+          name: interface_name,
+          major: major,
+          minor: minor,
+          exchanged_msgs: exchanged_msgs,
+          exchanged_bytes: exchanged_bytes
+        }
       end)
 
     # groups_map could be nil, default to empty map
@@ -87,6 +124,7 @@ defmodule Astarte.AppEngine.API.Device.DeviceStatus do
       credentials_inhibited: credentials_inhibited,
       total_received_msgs: total_received_msgs,
       total_received_bytes: total_received_bytes,
+      previous_interfaces: previous_interfaces,
       groups: groups
     }
   end