+++ apps/astarte_appengine_api/test/support/database_test_helper.exs
@@ -62,6 +62,7 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
         aliases map<ascii, varchar>,
         introspection map<ascii, int>,
         introspection_minor map<ascii, int>,
+        old_introspection map<frozen<tuple<ascii, int>>, int>,
         protocol_revision int,
         first_registration timestamp,
         credentials_secret ascii,
@@ -75,6 +76,8 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
         pending_empty_cache boolean,
         total_received_msgs bigint,
         total_received_bytes bigint,
+        exchanged_msgs_by_interface map<frozen<tuple<ascii, int>>, bigint>,
+        exchanged_bytes_by_interface map<frozen<tuple<ascii, int>>, bigint>,
         last_credentials_request_ip inet,
         last_seen_ip inet,
         groups map<text, timeuuid>,
@@ -94,7 +97,7 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
      device_id, aliases, connected, last_connection, last_disconnection,
      first_registration, first_credentials_request, last_seen_ip, last_credentials_request_ip,
      total_received_msgs, total_received_bytes, inhibit_credentials_request,
-     introspection, introspection_minor
+     introspection, introspection_minor, exchanged_msgs_by_interface, exchanged_bytes_by_interface
   )
   VALUES
     (
@@ -104,7 +107,8 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
       {'com.test.LCDMonitor' : 1, 'com.test.SimpleStreamTest' : 1,
        'com.example.TestObject': 1, 'com.example.PixelsConfiguration': 1},
       {'com.test.LCDMonitor' : 3, 'com.test.SimpleStreamTest' : 0,
-       'com.example.TestObject': 5, 'com.example.PixelsConfiguration': 0}
+       'com.example.TestObject': 5, 'com.example.PixelsConfiguration': 0},
+      :exchanged_msgs_by_interface, :exchanged_bytes_by_interface
     );
   """
 
@@ -460,14 +464,29 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
       |> DatabaseQuery.statement(@insert_device_statement)
 
     devices_list = [
-      {"f0VMRgIBAQAAAAAAAAAAAA", 4_500_000, %{"display_name" => "device_a"}},
-      {"olFkumNuZ_J0f_d6-8XCDg", 10, nil},
-      {"4UQbIokuRufdtbVZt9AsLg", 22, %{"display_name" => "device_b", "serial" => "1234"}},
-      {"aWag-VlVKC--1S-vfzZ9uQ", 0, %{"display_name" => "device_c"}},
-      {"DKxaeZ9LzUZLz7WPTTAEAA", 300, %{"display_name" => "device_d"}}
+      {"f0VMRgIBAQAAAAAAAAAAAA", 4_500_000,
+       %{
+         {"com.example.TestObject", 1} => 9300,
+         {"com.example.PixelsConfiguration", 1} => 4230,
+         {"com.test.LCDMonitor", 1} => 10,
+         {"com.test.LCDMonitor", 0} => 42
+       },
+       %{
+         {"com.example.TestObject", 1} => 2_000_000,
+         {"com.example.PixelsConfiguration", 1} => 2_010_000,
+         {"com.test.LCDMonitor", 1} => 3000,
+         {"com.test.LCDMonitor", 0} => 9000
+       }, %{"display_name" => "device_a"}},
+      {"olFkumNuZ_J0f_d6-8XCDg", 10, nil, nil, nil},
+      {"4UQbIokuRufdtbVZt9AsLg", 22, %{{"com.test.LCDMonitor", 1} => 4},
+       %{{"com.test.LCDMonitor", 1} => 16}, %{"display_name" => "device_b", "serial" => "1234"}},
+      {"aWag-VlVKC--1S-vfzZ9uQ", 0, %{}, %{}, %{"display_name" => "device_c"}},
+      {"DKxaeZ9LzUZLz7WPTTAEAA", 300, %{{"com.test.SimpleStreamTest", 1} => 9},
+       %{{"com.test.SimpleStreamTest", 1} => 250}, %{"display_name" => "device_d"}}
     ]
 
-    for {encoded_device_id, total_received_bytes, aliases} <- devices_list do
+    for {encoded_device_id, total_received_bytes, interface_msgs_map, interface_bytes_map,
+         aliases} <- devices_list do
       device_id = Base.url_decode64!(encoded_device_id, padding: false)
 
       insert_device_query =
@@ -475,6 +494,8 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
         |> DatabaseQuery.put(:device_id, device_id)
         |> DatabaseQuery.put(:aliases, aliases)
         |> DatabaseQuery.put(:total_received_bytes, total_received_bytes)
+        |> DatabaseQuery.put(:exchanged_msgs_by_interface, interface_msgs_map)
+        |> DatabaseQuery.put(:exchanged_bytes_by_interface, interface_bytes_map)
 
       DatabaseQuery.call!(client, insert_device_query)
 
@@ -488,6 +509,23 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
         DatabaseQuery.call!(client, insert_alias_query)
       end
     end
+
+    old_introspection_statement = """
+    INSERT INTO autotestrealm.devices
+    (device_id, old_introspection) VALUES (:device_id, :old_introspection)
+    """
+
+    old_introspection = %{{"com.test.LCDMonitor", 0} => 1}
+
+    {:ok, device_id} = Astarte.Core.Device.decode_device_id("f0VMRgIBAQAAAAAAAAAAAA")
+
+    query =
+      DatabaseQuery.new()
+      |> DatabaseQuery.statement(old_introspection_statement)
+      |> DatabaseQuery.put(:device_id, device_id)
+      |> DatabaseQuery.put(:old_introspection, old_introspection)
+
+    DatabaseQuery.call!(client, query)
 
     query =
       DatabaseQuery.new()